Improve GUI by rendering left and right panel with titles
Include stub class as dependency injection for Book so that CI/CD
pass. Labels and titles for left and right panel to follow our
Figma mockup.

Use of gridpane with row and column coordinates so that it is easy
to create and maintain the GUI main window. For example, assigning
coordinates (0,0) and (0,1) left and right panel titles while
controlling the row constraints will give us amber space for our
patron and book list panels to be rendered properly.