Package book: Add Book class and classes representing fields of Book
There is lack of objects and utility classes representing books in a
library.

Our application requires such representations to be able to
keep track and store information of books.

Let's introduce the following classes in a separate book package to help
store information regarding books:
* Book - Representing a book with name, isbn, authors and category tags
* Author - Representing an author of a book
* ISBN - Representing the isbn of a book
* BookName - Representing the name of a book
* Add separate test classes to unit test each of the classes above

Each field is represented as a class so that we abide to the Open-Closed
Principle in SOLID. For example, Author may only contain a String now,
but we may extend Author to include other attributes of the author in
future.

All these information is encapsulated in a separate book package so as
to abide to the encapsulation principle of OOP.