Update GUI components
Text in person and book cards will be truncated if too long, thus
wrapText is set to true to allow to prevent text truncation.

Certain parameters of the book command are optional such as the
author and the category tags. Therefore, there is a need to check
and ensure that the GUI renders these fields properly. In the
BookList.fxml, the visibility and managed attributes of the label
are set to false be default and only toggled to true in BookCard
if the author list supplied is not empty.

Different book tags have different colors and will change as accord
based on the status of the book.