Book package : Adapt Storage and Model
The small classes under Book package has already been adapted, but the
higher level classes concerning the overall Model and Storage is not
adapted to store Book object.

Let's,
* Create JsonAdaptedXYZ classes for XYZ = Book and Author to convert
Book and Author to Json objects
* Adapt Model and ModelManager so that we can perform CRUD operations
on a list of books in the model.
* Adapt AddressBook and ReadOnlyAddressBook to include Book
functionalities
* Adapt JsonSerializableAddressBook so that AddressBook can store
serialized JsonAdaptedBook data

Some unrelated test classes such as AddCommandTest are also modified
because they contain Stubs of interfaces modified above, and therefore
needs to implement the modifications in the stub.