Storage and Model Component: Minor Refactoring and Adding Tests
getTypicalAddressBook() is a method that is used for many test cases,
and is located inside TypicalPersons file.

It does not make sense to be inside that file because
getTypicalAddressBook() also needs to initialize the books, which are
inside TypicalBooks file.

Furthermore, the higher level Storage and Model components are yet to be
tested properly after the Book adaptation. Some test data json files are
also outdated.

Let's,
* Refactor getTypicalAddressBook() into a new TypicalAddressBook file,
and update all references to the method call in all test classes
* Update the test json data to include books
* Complete test classes for all Storage and Model components to provide
better test coverage

Updating test json data will prevent us the Storage component when
throwing error during testing. This is because the storage component
throws error when it tries to read book data from our sample data, but
the sample data does not contain "books" field.