Package Command: Create EditBookCommand Class
Our application does not include the logic for editing the book command, so that we can edit the information stored in the book.

There is a need to have such logic so that we can edit book objects in case a librarian enters something wrongly.

Let's include
* A EditBookCommand class which encapsulates the logic as to how the book object can be changed.
* A EditBookDescriptor class which encapsulates the parameters that will be changed.

This EditBookCommand class will be used by the EditBookCommandParser later.