Command: Made first step to refactor into packages
Our current parser does not support the parsing of commands with
2-word command words such as "patron add" or "book edit"

Let's:
* Add a layer of abstraction (i.e. BookCommandParser and
PatronCommandParser) so that parseCommand() from the original top-level
parser can call the parser() functions in these newly created classes
* Package all patron related commands into patron package, and book
related edit, delete, find, list, add commands into book package
* Package the same way for patron command parsers and book command
parsers
* Remove "COMMAND_WORD" from individual command classes, and refactor
them as a public static variable in Command class

Refactoring "COMMAND_WORD" is needed as some of the command words are
the same for different groups of commands (e.g. patron and book commands
). Refactoring eases the import, and reduces the chances of having
inconsistencies. It is also easier to change in the future, if needed.

The layer of abstraction and packaging is also needed so that we apply
SLAP principals, and avoid having long method for parseCommand()
function.