Book model: Add timeAdded attribute
Book object cannot be uniquely identified because according to our
software requirement, we can have multiple books of the same name, isbn,
author, and tags in our library.

This causes an issue when deleting books, because the inherent remove()
function used by JavaFX's ObservableList will only delete the first book
that is equal to the book we want to delete, based on equals(). Without
a unique timeAdded attribute, our equals() function will return true
for two books with the same name, isbn, tags and authors. This causes us
to always delete only the first book.

Let's:
* Use the time the book is added to our software to uniquely identify
the book
* Change test json data to include the new attribute

Using time added is a safe choice because we can assume no user can
add the same book to our system within the same millisecond. Users are
also not allowed to change the timeAdded attribute of the book using
Edit command.