Book : Update equals() and hashCode() methods to be not case sensitive and spacing sensitive
Book attributes such as Author and BookName are case and spacing
sensitive when two objects are compared. This is considered as a low
severity bug according to the module website.

Furthermore, books with the same authors but with different order of
authors are considered different. This can also be argued as a bug.

Let's
* Modify hashCode() and equals() methods of book attributes so that
the are not case and spacing sensitive for comparison