Book :  Add requesters attribute
Book class does not have attribute to store information of patrons
who are requesting for it.

This prevents us from implementing request command functionality to
fulfill our user story of keeping track of book requests.

Let's:
* Add a set of patrons as requester attribut to book
* Update sample data and test classes to reflect the update

Requesters are represented as a set because the order in which the
request happens should not matter for our use case.